+++ riverpie/lib/src/proxy_ref.dart
@@ -1,5 +1,6 @@
 import 'package:meta/meta.dart';
 import 'package:riverpie/src/action/dispatcher.dart';
+import 'package:riverpie/src/container.dart';
 import 'package:riverpie/src/labeled_reference.dart';
 import 'package:riverpie/src/notifier/base_notifier.dart';
 import 'package:riverpie/src/notifier/notifier_event.dart';
@@ -68,4 +69,7 @@ class ProxyRef implements Ref {
   void message(String message) {
     _ref.message(message);
   }
+
+  @override
+  RiverpieContainer get container => _ref.container;
 }